Major refactor of json serialization
I used to store in the json the path+line+column information as
a string. When loading back the information I would parse it to split
the various parts. I've moved to storing information as an array of
tables. This will allow me to add future features such as timestamps and
line content information.