Check current string at target loc in preview.
Previously we used the last modified time to check if the file had
changed since entries were added. This is not very useful because saving
a file changes the modified time, even if the file itself didn't change.
Instead we record the bytes offset of the target line and seek/read it
back to compare it against the stored info and color the preview
accordingly.
Also improve trimming of long lines to fit into preview window.